requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
Pillow==10.0.0
numpy==1.26.0
ST7789==0.0.1   # or use luma.lcd if your SPI display uses that driver



# Development and validation tools
pytest==7.4.2
pytest-mock==3.11.1
flake8==6.1.0
pylint==3.0.2